Building Information, Building Creation on Map and Soldier Creation were done.
There are barracks(blue) and powerplants(yellow) buildings on infinite scroll. These building can be dragged to map. Barrack size is 3x3 and PowerPlant size is 2x3. They can not be overlapped.

On Information part, building's name, image are shown. If barrack is shown on information page, soldier can be created with button. Soldier place starts front of the barrack and surrounds the building.

Grid model and controller added to project to manage the map.